Reframe skill/save prompt as "check prompt", make small refactors
- Use "check" instead of "skill/save" language in dialog
- Add perception
- Allow DC of 0
- Make all DOM ids globally unique
- Use foundry application API to setup tabs
- Keep dialog open after posting prompt
