+++ packs/macros/generate-check-prompt.json
@@ -1,9 +1,9 @@
 {
     "_id": "jHHOV1DdYytLO5VB",
     "author": "cXJZ4yzVgtPDwLCg",
-    "command": "game.pf2e.gm.skillSavePrompt();",
+    "command": "game.pf2e.gm.checkPrompt();",
     "img": "icons/svg/d20.svg",
-    "name": "Generate Skill/Save Prompt",
+    "name": "Generate Check Prompt",
     "ownership": {
         "default": 0
     },
