+++ src/global.ts
@@ -44,12 +44,12 @@ import type { PF2ECONFIG, StatusEffectIconTheme } from "@scripts/config/index.ts
 import type { DicePF2e } from "@scripts/dice.ts";
 import type {
     calculateXP,
+    checkPrompt,
     editPersistent,
     launchTravelSheet,
     perceptionForSelected,
     rollActionMacro,
     rollItemMacro,
-    skillSavePrompt,
     stealthForSelected,
     xpFromEncounter,
 } from "@scripts/macros/index.ts";
@@ -86,10 +86,10 @@ interface GamePF2e
         rollItemMacro: typeof rollItemMacro;
         gm: {
             calculateXP: typeof calculateXP;
+            checkPrompt: typeof checkPrompt;
             editPersistent: typeof editPersistent;
             launchTravelSheet: typeof launchTravelSheet;
             perceptionForSelected: typeof perceptionForSelected;
-            skillSavePrompt: typeof skillSavePrompt;
             stealthForSelected: typeof stealthForSelected;
             xpFromEncounter: typeof xpFromEncounter;
         };
