+++ src/scripts/macros/check-prompt/index.ts
@@ -1,0 +1,1 @@
+export { checkPrompt } from "./prompt.ts";
