--- src/scripts/macros/skill-save-prompt/index.ts
@@ -1,1 +1,0 @@
-export { skillSavePrompt } from "./skill-save-prompt.ts";
