+++ src/styles/gm/_index.scss
@@ -1,2 +1,2 @@
-@import "skill-save-prompt";
+@import "check-prompt";
 @import "travel-speed";
