+++ static/lang/en.json
@@ -353,19 +353,20 @@
             },
             "Party": {
                 "BlankSlate": "This party doesn't have any members. Drag a creature from the actor sidebar to get started.",
-                "SkillSavePrompt": {
+                "CheckPrompt": {
                     "AdjustDifficulty": "Adjust Difficulty",
                     "ChooseLores": "Choose Lores",
                     "ChooseSaves": "Choose Saves",
-                    "ChooseSkills": "Choose Skills",
+                    "ChooseSkills": "Choose Skills/Perception",
                     "LevelDC": "Level-Based DC",
                     "Post": "Generate Prompt",
                     "PromptTitle": "Prompt Title",
                     "SecretCheck": "Secret Check",
                     "SetDC": "Set DC",
                     "SimpleDC": "Simple DC",
-                    "Title": "Generate Skill/Save Prompt",
-                    "TitleShort": "Skill/Save Prompt"
+                    "SkillsPerception": "Skills/Perception",
+                    "Title": "Generate Check Prompt",
+                    "TitleShort": "Check Prompt"
                 },
                 "Coin": "Coin",
                 "ClearActivities": {
