+++ static/templates/actors/party/regions/exploration.hbs
@@ -46,8 +46,8 @@
         <header class="content-header">
             <a data-action="clear-exploration">{{localize "PF2E.Actor.Party.ClearActivities.Label"}}</a>
             <div class="buttons">
-                <button type="button" data-action="prompt" title="{{localize "PF2E.Actor.Party.SkillSavePrompt.TitleShort"}}"><i class="fa-solid fa-dice-d20"></i></button>
-                <button type="button" data-action="rest" title="{{localize "PF2E.Actor.Party.Rest"}}"><i class="fa-solid fa-fw fa-bed"></i></button>
+                <button type="button" data-action="prompt" data-tooltip="PF2E.Actor.Party.CheckPrompt.TitleShort"><i class="fa-solid fa-dice-d20"></i></button>
+                <button type="button" data-action="rest" data-tooltip="PF2E.Actor.Party.Rest" ><i class="fa-solid fa-fw fa-bed"></i></button>
             </div>
         </header>
     {{/if}}
