+++ static/templates/actors/party/sheet.hbs
@@ -2,7 +2,7 @@
     <header>
         <div class="frame-container">
             <div class="frame">
-                <img class="player-image" alt="{{actor.name}}" src="{{actor.img}}" title="{{actor.name}}" data-edit="img" />
+                <img class="player-image" alt="{{actor.name}}" src="{{actor.img}}" data-tooltip="{{actor.name}}" data-edit="img" />
             </div>
         </div>
         <div class="details">
@@ -77,7 +77,7 @@
                                 </div>
                                 <div class="item-controls">
                                     {{#if (and @root.options.editable (not item.temporary))}}
-                                        <a class="item-control item-delete" title="{{localize "PF2E.DeleteItemTitle"}}"><i class="fa-solid fa-fw fa-trash"></i></a>
+                                        <a class="item-control item-delete" data-tooltip="PF2E.DeleteItemTitle"><i class="fa-solid fa-fw fa-trash"></i></a>
                                     {{/if}}
                                 </div>
                             </li>
